CHROMIUM-only: Enable wake source for sdcard card detect pin
Currently the sdcard reader IRQ is not marked as a wake source in
the kernel. When the system is in S0ix inserting a sdcard causes
an interrupt. The kernel IRQ core discards the interrupt and the
interrupt status is not cleared by the sdcard reader IRQ handler.

This patch implements a workaround in the gpio pinctrl driver to
mark the parent IRQ of the sdcard as a wakeup source. So, the kernel
can now identify the wake source. And, the parent handler of sdcard
controller can clear the interrupt.

This patch sets IRQ 14 as a wake source by default for all GPIO's which
are controlled by GPIO pinctrl driver. This is a tentaive fix which
will be reverted once we get the correct fix.

BUG=b:36306698
TEST=Insert sdcard in S0ix and observe a wakeup to GUI.

Signed-off-by: Venkateswarlu Vinjamuri <venkateswarlu.v.vinjamuri@intel.com>

Change-Id: I687aabacf4dab4804896efabc86af88e21973ca1
Reviewed-on: https://chromium-review.googlesource.com/472126
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Tested-by: Venkateswarlu V Vinjamuri <venkateswarlu.v.vinjamuri@intel.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Venkateswarlu V Vinjamuri <venkateswarlu.v.vinjamuri@intel.com>