CHROMIUM: config: Enable camera on Poppy and friends
Enable all config entries necessary for camera subsystem and related
auxiliary devices (sensors, VCM) on Poppy and friends. All the
board-specific drivers are set to be modules, so that they do not add to
boot time of unrelated boards.

The split configs are normalized and unrelated config entries are
removed from the diff.

BUG=b:37020027
TEST=compile

Change-Id: I48140ca2e3565e1799edfa31a2b63cd1417419ba
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/500237
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>