CHROMIUM: arm64: dts: mt8176-rowan: configure the front camera
Additional controls are required to make the front camera working.
Including:
- Set the MUX to MIPI to USB path
- Set VGP4 to 2.8v and leavel it always-on
- Set the enable pin for VGP4

Since this camera is behind the MIPI to USB converter, the kernel will
recognize it as the UVC device. There is no proper way to control those
pins and the regulator in the UVC driver.

Add front_camera_control node to configure those pins and fix the
voltage to 2v8.

BUG=b:36465471
BRANCH=none
TEST= 1. Open the camera from Settings -> People -> Change picture
      2. Run powerd_dbus_suspend from console
      3. The camera resumes from suspend.

Change-Id: I15054095670c0ccbd48e359db3464b134eff154d
Signed-off-by: Yidi Lin <yidi.lin@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/476450
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>