UPSTREAM: af_key: Fix sadb_x_ipsecrequest parsing
The parsing of sadb_x_ipsecrequest is broken in a number of ways.
First of all we're not verifying sadb_x_ipsecrequest_len.  This
is needed when the structure carries addresses at the end.  Worse
we don't even look at the length when we parse those optional
addresses.

The migration code had similar parsing code that's better but
it also has some deficiencies.  The length is overcounted first
of all as it includes the header itself.  It also fails to check
the length before dereferencing the sa_family field.

This patch fixes those problems in parse_sockaddr_pair and then
uses it in parse_ipsecrequest.

BUG=b:64197102
TEST=Build and run

Change-Id: I0e0a0aaf30fb272b935913df1a51e5567019baf0
Reported-by: Andrey Konovalov <andreyknvl@google.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Steffen Klassert <steffen.klassert@secunet.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 096f41d3a8fc)
Reviewed-on: https://chromium-review.googlesource.com/749621
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>