CHROMIUM: iwl7000: mvm: use scnprintf() instead of snprintf()
It's safer to use scnprintf() here because the buffer might
be too short for the full format strings. In most cases
this isn't true because of external limits on the values.

In one case, this fixes a stack data leak.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: d7c3bdcc0b8605fea1d784528c5e3e91eef980aa