CHROMIUM: tpm: cr50_spi: reduce flow control timeout
Reduce the timeout waiting for flow control readiness from 5s
to 100ms. This value is used by firmware, and is more in line
with other timeouts defined by the spec.

Keeping a long timeout leads to unnecessary delays if cr50
for any reason misses the beginning of the transaction and
doesn't indicate readiness during the flow control phase.
Retrying after a smaller timeout is a better strategy.

BUG=b:37366209
TEST=boot, login on gru

Change-Id: I618d4225911d10b151385f7cad68afab6c0aa323
Signed-off-by: Andrey Pronin <apronin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/482182
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>