CHROMIUM: soc: rockchip: power-domain: Ignore 'shift-count-negative' warning
The following warning is generated when building with clang:

drivers/soc/rockchip/pm_domains.c:726:22: error: shift count is negative
  [-Werror,-Wshift-count-negative]
        [RK3399_PD_TCPD0]       = DOMAIN_RK3399(8, 8, -1, false),
                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
drivers/soc/rockchip/pm_domains.c:101:2: note: expanded from macro
  'DOMAIN_RK3399'
        DOMAIN(pwr, status, req, req, req, wakeup)
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
drivers/soc/rockchip/pm_domains.c:88:27: note: expanded from macro 'DOMAIN'
        .req_mask = (req >= 0) ? BIT(req) : 0,          \
                                 ^~~~~~~~
include/linux/bitops.h:6:24: note: expanded from macro 'BIT'

The BIT macro is evaluated with the negative value -1, even though the
resulting value would not be assigned. This is due to a clang bug:
https://bugs.llvm.org//show_bug.cgi?id=10030

A short-term fix in clang is not in sight and disabling the warning
completely because of this occurrence is not desirable. Instead disable
the warning locally just for the code in the subdirectory.

BUG=chromium:702741
TEST=build with clang for gru

Change-Id: Idc77b962d6080c3b11de4528205d53248de2bdb3
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/484689
Reviewed-by: Brian Norris <briannorris@chromium.org>