CHROMIUM: Remove duplicate declarations of sys_call_ptr_t
clang generates errors because sys_call_ptr_t is declared in different
locations. Remove the duplicates and include the header with the
declaration instead. Add the asmlinkage tag ("tells the compiler that
the function should not expect to find any of its arguments in registers
(a common optimization), but only on the CPU's stack") to the
declaration as upstream does and some of the dupes already did.

This is part of what 034042cc1e28 ("x86/entry/syscalls: Move
syscall table declarations into asm/syscalls.h") does. We don't do a
full backport since the change has further dependencies in which we
are not interested.

BUG=chromium:702746
TEST=build with clang for edgar and hana
  build with gcc for gale

Change-Id: Id690a6a0f29847766cefb64f2e5d699e210c6954
Reviewed-on: https://chromium-review.googlesource.com/492367
Commit-Ready: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>