CHROMIUM: mfd: cros-ec: bring in EC command definitions
We are adding support for new sensor api, so we need to bring in new
EC command definitions. To do that we are fetching ec_command.h at
<c5332f6fe7d4d5197ac2154319f5e7dd4436d62a> from the EC repository,
which contains canonical version of the file.

The new alignment changes brought in are not meant to change the
existing layout, but rather document it/make it explicit.

BUG=chromium:713814
TEST=Built/boot on caroline

Change-Id: I2161aa2c444d13420029515188e1dd82bdd1c52c
Signed-off-by: Nick Vaccaro <nvaccaro@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/484746
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>