FIXUP: CHROMIUM: drm/img-rogue: Separate the config option for DDK 1.7
If the old and the new driver versions are both enabled and built into the
kernel, builds fail with lots of duplicate symbols. If both are enabled
to be built as module, it is not well defined which of the drivers
will be loaded since the same symbols are exported by other drivers.
Building the old driver into the kernel and the new driver as module
doesn't do any good either since both versions use and export the same
symbols.

On top of that, defining a configuration option for the old version of the
driver and at the same time listing other configuration options as
depending on the new version doesn't do any good either for those other.

Restrict configuration options to ensure that test builds don't fail and
that only buildable configurations are possible.

Note that this doesn't guarantee that buildable options actually work;
exported symbols still overlap. Specifically, enabling both drivers results
in build warnings such as

WARNING: drivers/gpu/drm/img-rogue/1.7/pvrsrvkm:
	'DBGDrvGetServiceTable' exported twice. Previous export was in vmlinux

This commit does not attempt to address this problem.

While we are at it, add the driver version number into the Kconfig subject
for DRM_POWERVR_ROGUE_1_7.

BUG=b:37234148
TEST=Boot to UI on Elm

Change-Id: I1fd48a6ef89506f9fda2bcd17b507d5ece6d4097
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/493793
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>