FIXUP: CHROMIUM: drm/img-rogue: Add 1.10 IMG PowerVR Rogue driver
If the old and the new driver versions are both enabled and built into the
kernel, builds fail with lots of duplicate symbols. If both are enabled
to be built as module, it is not well defined which of the drivers
will be loaded since the same symbols are exported by other drivers.
Building the old driver into the kernel and the new driver as module
doesn't do any good either since both versions use and export the same
symbols.

On top of that, defining a configuration option for the old version of the
driver and at the same time listing other configuration options as
depending on the new version doesn't do any good either for those other.

Restrict configuration options to ensure that test builds don't fail and
that only buildable configurations are possible.

Note that this doesn't guarantee that buildable options actually work;
exported symbols still overlap. This commit does not attempt to address
this problem.

While at it, add driver version numbers into Kconfig subjects.

This patch is effectively a repeat of commit 4825845477b1 ("FIXUP:
CHROMIUM: drm/img-rogue: Separate the config option for DDK 1.7").

BUG=b:117489304
TEST=Build arm64:allyesconfig

Change-Id: I3141456d928bf90341f00bef00c718bc03e4739a
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1613898
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Zubin Mithra <zsm@chromium.org>