FIXUP: CHROMIUM: drm/img-rogue: Separate out config option for DDK 1.8 and 1.9
If the old and the new driver versions are both enabled and built into the
kernel, builds fail with lots of duplicate symbols. If both are enabled
to be built as module, it is not well defined which of the drivers
will be loaded since the same symbols are exported by other drivers.
Building the old driver into the kernel and the new driver as module
doesn't do any good either since both versions use and export the same
symbols.

Restrict configuration options to ensure that test builds don't fail and
that only buildable configurations are possible.

Note that this doesn't guarantee that buildable options actually work;
exported symbols may still overlap.

BUG=b:64932312
TEST=Boot to UI on Elm

Change-Id: I2bee37cfc35afd8fdfbbe10d9a5cd5f2ee5203cc
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/777120
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Frank Binns <frank.binns@imgtec.com>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>