UPSTREAM: drm/amdkfd: Add wavefront context save state retrieval ioctl
Wavefront context save data is of interest to userspace clients for
debugging static wavefront state. The MQD contains two parameters
required to parse the control stack and the control stack itself
is kept in the MQD from gfx9 onwards.

Add an ioctl to fetch the context save area and control stack offsets
and to copy the control stack to a userspace address if it is kept in
the MQD.

BUG=b:136019532
TEST=boot grunt, octopus, run login_LoginSuccess, cheets_StartAndroid

Signed-off-by: Jay Cornwall <Jay.Cornwall@amd.com>
Reviewed-by: Felix Kuehling <Felix.Kuehling@amd.com>
Signed-off-by: Felix Kuehling <Felix.Kuehling@amd.com>
Acked-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 5df099e8bc83f4f3af8711ee0b9b8faef359ffff)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: Ifc44926b773a47b5fe3e7e77265396844e72ba14