CHROMIUM: iwl7000: mvm: reset the HW before dumping if HW error is detected
If the hardware is stuck, we can't read any of the memory we need to
dump it, so we end up printing only 0xa5a5a5a5, which is useless.

To solve this, poke the hardware by triggering a reset and re-enabling
the clocks if we detect a HW error.

Additionally, for this to work, remove unnecessary bh protection when
calling the dump functions, so we can sleep to wait for the register
writes.

This is a combination of two patches from the iwl7000-tree.

BUG=b:37920902
TEST=Run continuous connect/disconnect cycles

Change-Id: If7eec4a60eee806ac63198832193c116b22697ba
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 054fad500f7398f18ea447778bc38f5361fedc5a
iwl7000-tree: 47e688654d273b8d7ba5ee963ba9a8a3674952de
Reviewed-on: https://chromium-review.googlesource.com/497770
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>