CHROMIUM: iwl7000: pcie: don't disable bh when handling FW errors
When we started using threaded irqs, all the opmode calls were changed
to be called with local_bh disabled.  The reason for this was it was
that mac80211 needs that.  When we are handling FW errors, mac80211 is
not involved, so we don't need it.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: fca084f1425872027d64817bf9c4602c328a06c0