CHROMIUM: iwl7000: mvm: print base HW address during init
It's sometimes hard to find out which HW address the iwlwifi device is
using, for instance when reading crouded sniffer logs.  To make it
easier, print out an info level message with the HW address as soon as
we know it.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: e8d5c9362bff90a751a4c30577d9ac53d1352350