CHROMIUM: iwl7000: pcie: delete the Tx queue timer earlier upon firmware crash
When the firmware crashes, the transmit queues can't make
any progress. This is why we stop the counter that monitor
the transmit queues' activity.
The call that notifies the error to the op_mode may take
a bit of time, so stop the timer of the transmit queues
earlier.

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 2d2aaabfabf0ea3061527c9fdb247f4653fe620b