CHROMIUM: low-mem: Add "available" sysfs file.
This allows consumers to determine how much memory is "available"
according to the memory pressure calculations performed in the kernel.

This reports available memory in megabytes to use the same units as the
"margin" file in the same directory.

BUG=chromium:722418
TEST=Booted kernel, verified available number looked sane

Signed-off-by: Justin TerAvest <teravest@chromium.org>
Change-Id: Ieb83e70607c0d549fc2291a87f07a5a5534c5c31
Reviewed-on: https://chromium-review.googlesource.com/506768
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
(cherry picked from commit e744f58098050a59dd870444c9392804a81ebc02)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>