CHROMIUM: iwl7000: chromeOS: include cfg80211
In order to simplify maintenance of the iwl7000 driver, include a
copy of cfg80211. This copy is renamed to iwl7000_cfg80211 and, just
like we already had with mac80211, all symbols within it are renamed
so that the regular cfg80211 and this one can be installed at the
same time, and the module loader will pick the correct one at runtime.

This significantly reduces the amount of divergence/patching of the
iwl7000 driver from the regular mainline/Intel-internal driver, and
the amount of "glue code" needed. More new glue code is added to make
cfg80211 work on older kernels, but that code is mostly taken from the
backports project and well-tested already.

This does, however, remove the ability to have multiple wireless NICs
in use at the same time.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 95d76605d09a9d4f4937591ce9481d140c1a5bfc