CHROMIUM: [media] v4l2-core: Use kvmalloc() for potentially big allocations
There are multiple places where arrays or otherwise variable sized
buffer are allocated through V4L2 core code, including things like
controls, memory pages, staging buffers for ioctls and so on. Such
allocations can potentially require an order > 0 allocation from the
page allocator, which is not guaranteed to be fulfilled and is likely to
fail on a system with severe memory fragmentation (e.g. a system with
very long uptime).

Since the memory being allocated is intended to be used by the CPU
exclusively, we can consider using vmalloc() as a fallback and this is
exactly what the recently merged kvmalloc() helpers do. A kmalloc() call
is still attempted, even for order > 0 allocations, but it is done
with __GFP_NORETRY and __GFP_NOWARN, with expectation of failing if
requested memory is not available instantly. Only then the vmalloc()
fallback is used. This should give us fast and more reliable allocations
even on systems with higher memory pressure and/or more fragmentation,
while still retaining the same performance level on systems not
suffering from such conditions.

While at it, replace explicit array size calculations on changed
allocations with kvmalloc_array().

(An equivalent patch was sent upstream separately due to a big number of
merge conflicts: https://patchwork.linuxtv.org/patch/41547/)

BUG=chromium:721786
TEST=test_that video_VideoDecodeAccelerator.vp8 on veyron_minnie
TEST=test_that video_VideoEncodeAccelerator.vp8 on veyron_minnie

Change-Id: I690bd14bc15e60187bae6058bc2d7c91dcf232e5
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/509090
Reviewed-by: Pawel Osciak <posciak@chromium.org>