CHROMIUM: drm/img-rogue: generate WARNING on PVR_K
Call OSWarnOn() instead of OSDumpStack() when making a Debug Dump.

The DDK Debug Dump contains information useful for diagnosing problems.
It is produced automatically in a number of different scenarios. It
usually contains a stack dump.

By using OSWarnOn() - WARN_ON() - the stack dump is presented in a
standard way and can be used to trigger the ChromeOS crash reporter.

BUG=b:37098430
TEST=Boot to UI

Change-Id: Ia541b834c6917d1a97c75b84c42d287f154584a8
Signed-off-by: Afar Lin <afar.lin@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/512082
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>