CHROMIUM: [media] rockchip-vpu: Fix inactive reference frame configuration
If we assign 0 to the inactive DPB position in RPS packet,
HW will treat that as a true reference frame with pic_num 0
when frame pic_num 0 is in the reference list and cause
output picture corruption.

To fix it, this patch changes the code to assign an invalid
pic_num to RPS packet to prevent HW from referring to an
invalid reference frame.

BUG=b:36768258
TEST=Play the videos mentioned in the bug on Kevin

Change-Id: I5bb080cdb04e9f63db3c70fc47dc293333a7ac73
Signed-off-by: Alpha Lin <alpha.lin@rock-chips.com>
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/512426
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Alpha Lin <Alpha.Lin@rock-chips.com>