HACK: Input: elants_i2c - clear interrupt at resume
If a device is using edge-triggered interrupt, then, when it gets
enabled with enable_irq(), and edge came while interrupt was disabled,
the kernel is supposed to replay the interrupt.

Unfortunately variety of Strago devices do not properly describe IRQs in
their firmwares and replay does not work properly. As a workaround,
until we fix the firmware, let's read and discard the data from the
device, so interrupt is cleared before it is re-enabled.

BUG=b:62677107
TEST=Suspend/resume on Kefka, ensure touchscreen works.

Change-Id: Ic38ab11f842dec2c12de71593113f89c120ad839
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/517282
Reviewed-by: Benson Leung <bleung@chromium.org>
(cherry picked from commit f495ada03ddc7b546850cd0d0f9b29baf364c02e)
Reviewed-on: https://chromium-review.googlesource.com/539858
Reviewed-by: Guenter Roeck <groeck@chromium.org>