CHROMIUM: android: binder: another use of binder_lock_interruptible
When flushing/killing threads there is chance that we just fetched
data from userspace and now waiting on the binder lock to be able to
select a new thread and start submitting request. Convert it to
binder_lock_interruptible(), so if the thread is selected by the OOM
killer it will die quickly.

BUG=chromium:718637
TEST=Boot on Caroline, play with Android apps

Change-Id: Iff6126fea31f8db17255f368b95e1902ba90aaa9
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/522233
Reviewed-by: Douglas Anderson <dianders@chromium.org>